add DESTDIR to Makefile
This patch adds the DESTDIR variable to the Makefile to specify the
installation location. This variable is very useful for packaging.

Currently, only 'prefix' variable is supported in Makefile, so if you
specify '/usr' to 'prefix', it is difficult to package the result of
the build separately.

If you use the DESTDIR, even if 'prefix' is '/usr', you can install
it in a separate location via DESTDIR, then import the directory and
you can package the result neatly.

Signed-off-by: Inho Oh <inho.oh@sk.com>